Break up test cases
