+++ subprojects/dependency-lock/src/integTest/groovy/org/gradle/internal/dependencylock/generation/DependencyLockFileResolvedDependenciesIntegrationTest.groovy
@@ -14,58 +14,11 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.dependencylock
+package org.gradle.internal.dependencylock.generation
 
-import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.test.fixtures.file.TestFile
-
-import static org.gradle.initialization.StartParameterBuildOptions.DependencyLockOption
 import static org.gradle.internal.dependencylock.fixtures.DependencyLockFixture.*
 
-class DependencyLockFileGenerationIntegrationTest extends AbstractIntegrationSpec {
-
-    private static final String MYCONF_CUSTOM_CONFIGURATION = 'myConf'
-
-    TestFile lockFile
-    TestFile sha1File
-
-    def setup() {
-        lockFile = file('gradle/dependencies.lock')
-        sha1File = file('gradle/dependencies.lock.sha1')
-    }
-
-    def "does not write lock file if no dependencies were resolved"() {
-        given:
-        buildFile << mavenRepository(mavenRepo)
-        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
-        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        !lockFile.exists()
-        !sha1File.exists()
-    }
-
-    def "does not write lock file if all dependencies failed to be resolved"() {
-        given:
-        buildFile << mavenRepository(mavenRepo)
-        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
-        buildFile << """
-            dependencies {
-                myConf 'foo:bar:1.5'
-            }
-        """
-        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        when:
-        failsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        !lockFile.exists()
-        !sha1File.exists()
-    }
+class DependencyLockFileResolvedDependenciesIntegrationTest extends AbstractDependencyLockFileGenerationIntegrationTest {
 
     def "can create locks for dependencies with concrete version"() {
         given:
@@ -480,348 +433,5 @@ ${commentAndLockFileVersionJson()}
   ]
 }"""
         sha1File.text == '60509fcde47220bc2bbd808a5ecb8a5839fe323c'
-    }
-
-    def "can create locks for multi-project builds"() {
-        given:
-        mavenRepo.module('my', 'dep', '1.5').publish()
-        mavenRepo.module('foo', 'bar', '2.3.1').publish()
-        mavenRepo.module('other', 'company', '5.2').publish()
-
-        buildFile << """
-            subprojects {
-                ${mavenRepository(mavenRepo)}
-            }
-
-            project(':a') {
-                ${customConfigurations('conf1')}
-    
-                dependencies {
-                    conf1 'my:dep:1.5'
-                }
-
-                ${copyLibsTask('conf1')}
-            }
-
-            project(':b') {
-                ${customConfigurations('conf2')}
-
-                dependencies {
-                    conf2 'foo:bar:2.3.1'
-                }
-
-                ${copyLibsTask('conf2')}
-            }
-
-            project(':c') {
-                ${customConfigurations('conf3')}
-
-                dependencies {
-                    conf3 'other:company:5.2'
-                }
-
-                ${copyLibsTask('conf3')}
-            }
-        """
-        settingsFile << "include 'a', 'b', 'c'"
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":a",
-      "configurations": [
-        {
-          "name": "conf1",
-          "dependencies": [
-            {
-              "moduleId": "my:dep",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "path": ":b",
-      "configurations": [
-        {
-          "name": "conf2",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "2.3.1",
-              "lockedVersion": "2.3.1"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "path": ":c",
-      "configurations": [
-        {
-          "name": "conf3",
-          "dependencies": [
-            {
-              "moduleId": "other:company",
-              "requestedVersion": "5.2",
-              "lockedVersion": "5.2"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == 'ba328f04e6f66725791db639c04f9b09fe0b69da'
-    }
-
-    def "subsequent builds do not recreate lock file for unchanged dependencies"() {
-        given:
-        mavenRepo.module('foo', 'bar', '1.5').publish()
-
-        buildFile << mavenRepository(mavenRepo)
-        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
-        buildFile << """
-            dependencies {
-                myConf 'foo:bar:1.5'
-            }
-        """
-        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTasksExecuted(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '47ec5ad9e745cef18cea6adc42e4be3624572c9f'
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTaskSkipped(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '47ec5ad9e745cef18cea6adc42e4be3624572c9f'
-    }
-
-    def "recreates lock file for newly declared and resolved dependencies"() {
-        given:
-        mavenRepo.module('foo', 'bar', '1.5').publish()
-        mavenRepo.module('org', 'gradle', '7.5').publish()
-
-        buildFile << mavenRepository(mavenRepo)
-        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
-        buildFile << """
-            dependencies {
-                myConf 'foo:bar:1.5'
-            }
-        """
-        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTasksExecuted(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '47ec5ad9e745cef18cea6adc42e4be3624572c9f'
-
-        when:
-        buildFile << """
-            dependencies {
-                myConf 'org:gradle:7.5'
-            }
-        """
-
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTasksExecuted(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            },
-            {
-              "moduleId": "org:gradle",
-              "requestedVersion": "7.5",
-              "lockedVersion": "7.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '972a625f8f26cb15c53a9962cdbc9230ae551aec'
-    }
-
-    def "recreates lock file for removed, resolved dependencies"() {
-        given:
-        mavenRepo.module('foo', 'bar', '1.5').publish()
-        mavenRepo.module('org', 'gradle', '7.5').publish()
-
-        buildFile << mavenRepository(mavenRepo)
-        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
-        buildFile << """
-            dependencies {
-                myConf 'foo:bar:1.5'
-                myConf 'org:gradle:7.5'
-            }
-        """
-        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        when:
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTasksExecuted(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            },
-            {
-              "moduleId": "org:gradle",
-              "requestedVersion": "7.5",
-              "lockedVersion": "7.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '972a625f8f26cb15c53a9962cdbc9230ae551aec'
-
-        when:
-        buildFile.text = mavenRepository(mavenRepo) + customConfigurations(MYCONF_CUSTOM_CONFIGURATION) + """
-            dependencies {
-                myConf 'foo:bar:1.5'
-            }
-        """ + copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
-
-        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
-
-        then:
-        result.assertTasksExecuted(COPY_LIBS_TASK_PATH)
-        lockFile.text == """{
-${commentAndLockFileVersionJson()}
-  "projects": [
-    {
-      "path": ":",
-      "configurations": [
-        {
-          "name": "myConf",
-          "dependencies": [
-            {
-              "moduleId": "foo:bar",
-              "requestedVersion": "1.5",
-              "lockedVersion": "1.5"
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}"""
-        sha1File.text == '47ec5ad9e745cef18cea6adc42e4be3624572c9f'
-    }
-
-    private void succeedsWithEnabledDependencyLocking(String... tasks) {
-        withEnabledDependencyLocking()
-        succeeds(tasks)
-    }
-
-    private void failsWithEnabledDependencyLocking(String... tasks) {
-        withEnabledDependencyLocking()
-        fails(tasks)
-    }
-
-    private void withEnabledDependencyLocking() {
-        args("--$DependencyLockOption.LONG_OPTION")
     }
 }
