+++ subprojects/dependency-lock/src/integTest/groovy/org/gradle/internal/dependencylock/generation/DependencyLockFileUnresolvedDependenciesIntegrationTest.groovy
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.dependencylock.generation
+
+import static org.gradle.internal.dependencylock.fixtures.DependencyLockFixture.*
+
+class DependencyLockFileUnresolvedDependenciesIntegrationTest extends AbstractDependencyLockFileGenerationIntegrationTest {
+
+    def "does not write lock file if no dependencies were resolved"() {
+        given:
+        buildFile << mavenRepository(mavenRepo)
+        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
+        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
+
+        when:
+        succeedsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
+
+        then:
+        !lockFile.exists()
+        !sha1File.exists()
+    }
+
+    def "does not write lock file if all dependencies failed to be resolved"() {
+        given:
+        buildFile << mavenRepository(mavenRepo)
+        buildFile << customConfigurations(MYCONF_CUSTOM_CONFIGURATION)
+        buildFile << """
+            dependencies {
+                myConf 'foo:bar:1.5'
+            }
+        """
+        buildFile << copyLibsTask(MYCONF_CUSTOM_CONFIGURATION)
+
+        when:
+        failsWithEnabledDependencyLocking(COPY_LIBS_TASK_NAME)
+
+        then:
+        !lockFile.exists()
+        !sha1File.exists()
+    }
+}
